refactor: split Development panel connectivity to satisfy economy file size (#4175)
Refs #4175
